requirements.txt
python-dotenv==1.0.0
pillow==9.5.0
pyttsx3==2.90
flask==2.3.3
google-generativeai==0.3.1
pytesseract==0.3.10
langchain==0.1.0
langchain-openai==0.0.2
openai==1.3.0 